pyproject.toml
[project]
name = "retail-warehouse-scraper"
version = "0.1.0"
description = "AI agent for scraping retail and warehouse business prospect information"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pandas>=2.1.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.7.3",
    "pydantic-ai>=0.2.12",
    "sqlalchemy>=2.0.41",
    "httpx>=0.28.1",
    "openai>=1.83.0",
    "anthropic>=0.52.2",
    "firecrawl-py>=2.7.1",
    "structlog>=25.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
